iOS wizard: determine version from either opts or metadata file
